+++ bot/src/module/song.rs
@@ -481,18 +481,6 @@ impl command::Handler for Handler {
                 ctx.check_moderator()?;
                 self.player.pause()?;
             }
-            Some("detach") => {
-                ctx.check_moderator()?;
-
-                match self.player.clear_device() {
-                    Some(_) => {
-                        ctx.respond("Detached player");
-                    }
-                    None => {
-                        ctx.respond("No player configured");
-                    }
-                }
-            }
             Some("length") => {
                 let (count, duration) = self.player.length();
 
@@ -678,6 +666,9 @@ fn player_feedback_loop(
 
             move |e| {
                 match e {
+                    player::Event::Detached => {
+                        sender.privmsg(channel.as_str(), "Player is detached!");
+                    }
                     player::Event::Playing(echo, item) => {
                         if !echo || !*chat_feedback.read() {
                             return Ok(());