+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Setting for controlling player feedback in chat (`player/chat-feedback`).
 - Setting for controlling overlay update interval (`player/song-update-interval`).
 - `!admin settings` for reading and writing settings through chat.
-- `!song detach` to detach the current player, preventing it from playing back until it has been attached again. ([#27])
+- Introduced the setting `player/detached` to detach the player. ([#27])
 - Confirmation response when performing `!song open` and `!song close`. ([#36])
 - `!admin push <key> <value>` to insert values into settings which are collections.
 - `!admin delete <key> <value>` to delete values from settings which are collections.